+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-26T09:46:50.254Z")
+    updated: Optional[datetime] = parse_datetime("2026-08-26T11:10:52.849Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")